% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{aggregateWVMat}
\alias{aggregateWVMat}
\title{get a list of w.mat (enhenced by v.mat) and v.mat (enhenced by w.mat),
both aggregated for multiple SCTGs.}
\arguments{
\item{weighTable}{weight table}

\item{valueTable}{value table}

\item{SCTGs_}{SCTG numbers}

\item{origins}{labels}
}
\value{
a list of weight matrix and value matrix
}
\description{
get a list of w.mat (enhenced by v.mat) and v.mat (enhenced by w.mat),
both aggregated for multiple SCTGs.
}

