% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{analysis}
\alias{analysis}
\title{analysis matrix}
\arguments{
\item{mat}{matrix}

\item{sorted.origins}{labels}

\item{result.name}{path to make the retults}

\item{index.name}{path to make index file}

\item{matrix.name}{path to output the matrix}
}
\description{
analysis matrix
}

