% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{calculateWVMat}
\alias{calculateWVMat}
\title{get a list of w.mat (enhenced by v.mat) and v.mat (enhenced by w.mat).}
\arguments{
\item{weighTable}{weight table}

\item{valueTable}{value table}

\item{SCTG}{SCTG number}

\item{origins}{labels}
}
\value{
a list of weight matrix and value matrix
}
\description{
get a list of w.mat (enhenced by v.mat) and v.mat (enhenced by w.mat).
}

