% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{generateCircos}
\alias{generateCircos}
\title{generate circos input file}
\arguments{
\item{mat}{matrix}

\item{sorted.origins}{labels}

\item{SCTGs_}{SCTG numbers}

\item{origins}{labels}

\item{VWCTable}{matrix}
}
\description{
generate circos input file
}

