% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getNetworkPearson}
\alias{getNetworkPearson}
\title{get network pearson coefficient}
\arguments{
\item{mat}{matrix}

\item{sorted.origins}{labels}

\item{SCTGs_}{SCTG numbers}

\item{origins}{labels}

\item{VWCTable}{matrix}
}
\description{
get network pearson coefficient
}

