% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getVirtualMatrix}
\alias{getVirtualMatrix}
\title{calculate virtual water content matrix}
\arguments{
\item{weighTable}{weight table}

\item{valueTable}{value table}

\item{SCTGs_}{SCTG numbers}

\item{origins}{labels}

\item{VWCTable}{virtual water content table. It must have 51 row, from Alabama to Wyoming.
Columns have to be corresponding to SCTGs_.}
}
\value{
matrix
}
\description{
calculate virtual water content matrix
}

