% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{heatMap}
\alias{heatMap}
\title{make heat meaps valued on values provided for each state}
\arguments{
\item{values}{a vector indicate the value in each state}

\item{origins}{labels for the map}

\item{color}{choose among g, b, y, v. o}

\item{filename}{a path to output pdf}
}
\description{
make heat meaps valued on values provided for each state
}

