% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{new}
\alias{new}
\title{constructor}
\arguments{
\item{weighTable}{weight table}

\item{valueTable}{value table}

\item{SCTGs_}{SCTG numbers}

\item{origins}{labels}

\item{VWCTable}{matrix}

\item{dir}{path to VWCCrop.csv file}
}
\description{
constructor

constructor
}

