% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{toMatrix}
\alias{toMatrix}
\title{given weight or value option, SCTG, return corresponding matrix}
\arguments{
\item{tab}{weigh table or value table}

\item{SCTG}{SCTG number}
}
\value{
trade matrix
}
\description{
given weight or value option, SCTG, return corresponding matrix
}

