% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{TradeMatrix-class}
\alias{TradeMatrix}
\alias{TradeMatrix-class}
\title{class definition of TradeMatrix}
\description{
class definition of TradeMatrix
}
\section{Fields}{

\describe{
\item{\code{weightMatrix}}{a trade graph represented by matrix, enhanced by value matrix}

\item{\code{valueMatrix}}{a trade graph represented by matrix, enhanced by weight matrix}
}}
\section{Methods}{

\describe{
\item{\code{analysis(mat, sorted.origins, result.name = "data\\\\cropResult.csv",
  index.name = "data\\\\indexes.csv",
  matrix.name = "data\\\\cropMatrix.csv")}}{string doc is working}
}}

