% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{TradeTable-class}
\alias{TradeTable}
\alias{TradeTable-class}
\title{class definition of TradeTable}
\description{
class definition of TradeTable
}
\section{Fields}{

\describe{
\item{\code{filename}}{filename of input file}
}}

