% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{uscropnetwork}
\alias{uscropnetwork}
\alias{uscropnetwork-package}
\title{Analysis U.S. domestic crop trade network, in terms of weight, and vritual water
content.}
\description{
\enumerate{
 \item Data preprocessing.
 \item Virtual water content calculation.
 \item Network analysis.
}
}

